hsta_project/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
ci-info==0.2.0
click==8.0.1
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
decorator==4.4.2
Django==3.2
django-restframework==0.0.1
djangorestframework==3.12.4
docx==0.2.4
et-xmlfile==1.1.0
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
future==0.18.2
httplib2==0.19.1
idna==3.2
importlib-metadata==4.8.1
isodate==0.6.0
lxml==4.6.3
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.1
numpy==1.19.5
openpyxl==3.0.7
packaging==21.0
pandas==0.25.3
pathlib==1.0.1
Pillow==8.3.2
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.16.14
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
requests==2.26.0
scipy==1.5.4
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
traits==6.2.0
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
